ai-service/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.3
python-multipart==0.0.9
spacy==3.7.4
pdfminer.six==20231228
python-docx==1.1.0
scikit-learn==1.4.1.post1
numpy==1.26.4
pandas==2.2.1
pydantic-settings==2.2.1
